aegis-decision-service/requirements.txt
# A.E.G.I.S. Decision Service - Production Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# IBM watsonx.ai
ibm-watsonx-ai>=1.1.0

# HTTP Client for Langflow integration
requests>=2.31.0

# YAML support for OpenAPI export
pyyaml>=6.0.1

# Validation
pydantic>=2.5.0

# Environment variables
python-dotenv>=1.0.0

# Testing (dev dependencies, but included for convenience)
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.26.0  # For TestClient
